Загрузка файлов если админ, управление эвентами и новостями.

+++ main.py
@@ -8,7 +8,7 @@ import datetime
 from data.images import Images
 from data.videos import Videos
 from data.audio import Audios
-from data.classes import RegistrationForm, LoginForm, UploadForm
+from data.classes import RegistrationForm, LoginForm, UploadFormImage, UploadFormAudio, UploadFormVideo
 from werkzeug.security import generate_password_hash, check_password_hash
 import datetime
 from flask_login import LoginManager, login_user, current_user, login_required, logout_user
@@ -38,6 +38,11 @@ def get_registred_image():
     else:
         return "billy.jpg"
 
+def get_is_admin():
+    if current_user.is_authenticated:
+        return current_user.is_admin
+    else:
+        return False
 
 @login_manager.user_loader
 def load_user(user_id):
@@ -48,24 +53,20 @@ def load_user(user_id):
 @app.route('/')
 @app.route('/index')
 def index():
-    try:
-        events = open('data/actual_events.json', encoding='utf-8')
-        f = events.read()
-        afisha_data = json.loads(f)
-        na = open('data/news_articles.json', encoding='utf-8')
-        f1 = na.read()
-        NA = json.loads(f1)
-        session = db_session.create_session()
-        for event in afisha_data:
-            image = session.query(Images).filter(Images.id == event['image']).first()
-            event['image'] = f"static/images/{image.file_name}"
-        print(get_registred_image())
-        return render_template('./index.html',
-                               afisha=afisha_data,
-                               news=NA, not_registered=get_registred_id(), profile_image=get_registred_image())
-    except Exception as err:
-        return render_template("error.html", err=err, not_registered=get_registred_id(),
-                               profile_image=get_registred_image())
+    events = open('data/actual_events.json', encoding='utf-8')
+    f = events.read()
+    afisha_data = json.loads(f)
+    na = open('data/news_articles.json', encoding='utf-8')
+    f1 = na.read()
+    NA = json.loads(f1)
+    session = db_session.create_session()
+    for event in afisha_data:
+        image = session.query(Images).filter(Images.id == event['image']).first()
+        event['image'] = f"static/images/{image.file_name}"
+    return render_template('./index.html',
+                           afisha=afisha_data,
+                           news=NA, not_registered=get_registred_id(), profile_image=get_registred_image(),
+                           is_admin=get_is_admin())
 
 
 @app.route('/support')
@@ -165,7 +166,7 @@ def profile(profile_id):
         return render_template('profile.html', name=user.username, not_registered=get_registred_id(),
                                user_information=data, profile_image=image_profile, image_information=data1,
                                video_information=data2, audio_information=data3, this_user=this_user, user_id=profile_id,
-                               is_curr_user_admin=current_user.is_admin)
+                               is_curr_user_admin=get_is_admin())
     else:
         return render_template('nt_exist.html', id=profile_id, type="Пользователя",
                                not_registered=get_registred_id(), profile_image=get_registred_image())
@@ -265,7 +266,7 @@ def delete_from(redirto, typ, del_id):
     if typ == "image":
         image = session.query(Images).filter(Images.id == del_id).first()
         if image:
-            if (image.author_id == current_user.id) or current_user.is_admin:
+            if (image.author_id == current_user.id) or get_is_admin():
                 image.operation_type = "DELETED"
                 try:
                     os.remove('static/images/' + image.file_name)
@@ -274,7 +275,7 @@ def delete_from(redirto, typ, del_id):
     if typ == "video":
         video = session.query(Videos).filter(Videos.id == del_id).first()
         if video:
-            if (video.author_id == current_user.id) or current_user.is_admin:
+            if (video.author_id == current_user.id) or get_is_admin():
                 video.operation_type = "DELETED"
                 try:
                     os.remove('static/videos/' + video.file_name)
@@ -283,7 +284,7 @@ def delete_from(redirto, typ, del_id):
     if typ == "audio":
         audio = session.query(Audios).filter(Audios.id == del_id).first()
         if audio:
-            if (audio.author_id == current_user.id) or current_user.is_admin:
+            if (audio.author_id == current_user.id) or get_is_admin():
                 audio.operation_type = "DELETED"
                 try:
                     os.remove('static/audio/' + video.file_name)
@@ -302,37 +303,43 @@ def delete_from(redirto, typ, del_id):
 @login_required
 def load_image(type):
     try:
-        form = UploadForm()
+        form = UploadFormImage()
         errors = []
         session = db_session.create_session()
         form_title = ""
         if type == "prf":
             form_title = "профиля"
+        if type == "adm":
+            if get_is_admin():
+                form_title = "использования в материалах сайта"
+            else:
+                return redirect("/")
         if form.validate_on_submit():
-            print(form.file.data.filename)
             images = session.query(Images).filter(Images.file_name == form.file.data.filename,
                                                   Images.operation_type == "Exist").first()
             if not images:
                 form.file.data.save('static/images/' + form.file.data.filename)
                 user = session.query(Users).filter(Users.id == current_user.id).first()
                 if user:
-                    if user.profile_image:
-                        image = session.query(Images).filter(Images.file_name == user.profile_image).first()
-                        if image:
-                            image.operation_type = "DELETED"
-                            try:
-                                os.remove('static/images/' + user.profile_image)
-                            except Exception as err:
-                                print(err)
+                    if type == "prf":
+                        if user.profile_image and not get_is_admin():
+                            image = session.query(Images).filter(Images.file_name == user.profile_image).first()
+                            if image:
+                                image.operation_type = "DELETED"
+                                try:
+                                    os.remove('static/images/' + user.profile_image)
+                                except Exception as err:
+                                    print(err)
                     last_image = session.query(Images).order_by(Images.id.desc()).first()
                     image = Images()
                     image.id = last_image.id + 1
                     image.file_name = form.file.data.filename
-                    image.operation_type = "Exist"
+                    image.operation_type = "EXIST"
                     image.date = (str(datetime.date.today()) + " " + str(datetime.datetime.now().strftime("%H:%M:%S")))
                     image.author_id = user.id
                     session.merge(image)
-                    user.profile_image = form.file.data.filename
+                    if type == "prf":
+                        user.profile_image = form.file.data.filename
                     session.commit()
                     return redirect("/my_profile")
             else:
@@ -343,6 +350,105 @@ def load_image(type):
     except Exception as err:
         return render_template("error.html", err=err, not_registered=get_registred_id(),
                                profile_image=get_registred_image())
+
+@app.route('/load/video', methods=['GET', 'POST'])
+@login_required
+def load_video():
+    try:
+        form = UploadFormVideo()
+        errors = []
+        session = db_session.create_session()
+        if not get_is_admin():
+            return redirect("/")
+        if form.validate_on_submit():
+            videos = session.query(Videos).filter(Videos.file_name == form.file.data.filename,
+                                                  Images.operation_type == "Exist").first()
+            if not videos:
+                form.file.data.save('static/videos/' + form.file.data.filename)
+                user = session.query(Users).filter(Users.id == current_user.id).first()
+                if user:
+                    last_video = session.query(Videos).order_by(Videos.id.desc()).first()
+                    video = Videos()
+                    video.id = last_video.id + 1
+                    video.file_name = form.file.data.filename
+                    video.operation_type = "EXIST"
+                    video.date = (str(datetime.date.today()) + " " + str(datetime.datetime.now().strftime("%H:%M:%S")))
+                    video.author_id = user.id
+                    session.merge(video)
+                    session.commit()
+                    return redirect("/my_profile")
+            else:
+                errors.append("Видео с таким именем уже существует.")
+
+        return render_template("load_video_audio.html", form=form, additional_errors=errors,
+                               not_registered=get_registred_id(), profile_image=get_registred_image(),
+                               title='видео')
+    except Exception as err:
+        return render_template("error.html", err=err, not_registered=get_registred_id(),
+                               profile_image=get_registred_image())
+
+
+@app.route('/load/audio', methods=['GET', 'POST'])
+@login_required
+def load_audio():
+    try:
+        form = UploadFormAudio()
+        errors = []
+        session = db_session.create_session()
+        if not get_is_admin():
+            return redirect("/")
+        if form.validate_on_submit():
+            audios = session.query(Audios).filter(Audios.file_name == form.file.data.filename,
+                                                  Images.operation_type == "Exist").first()
+            if not audios:
+                form.file.data.save('static/audio/' + form.file.data.filename)
+                user = session.query(Users).filter(Users.id == current_user.id).first()
+                if user:
+                    last_audio = session.query(Audios).order_by(Audios.id.desc()).first()
+                    audio = Audios()
+                    audio.id = last_audio.id + 1
+                    audio.file_name = form.file.data.filename
+                    audio.operation_type = "EXIST"
+                    audio.date = (str(datetime.date.today()) + " " + str(datetime.datetime.now().strftime("%H:%M:%S")))
+                    audio.author_id = user.id
+                    session.merge(audio)
+                    session.commit()
+                    return redirect("/my_profile")
+            else:
+                errors.append("Аудио с таким именем уже существует.")
+
+        return render_template("load_video_audio.html", form=form, additional_errors=errors,
+                               not_registered=get_registred_id(), profile_image=get_registred_image(),
+                               title="аудио")
+    except Exception as err:
+        return render_template("error.html", err=err, not_registered=get_registred_id(),
+                               profile_image=get_registred_image())
+
+@app.route("/move_to_archive_<typ>/<int:id>")
+@login_required
+def move_to_archive(typ, id):
+    if typ == "event":
+        events = open('data/actual_events.json', encoding='utf-8')
+        f = events.read()
+        afisha_data = json.loads(f)
+        for k in afisha_data:
+            if id == k['id']:
+                k['is_in_archive'] = True
+        events.close()
+        events = open('data/actual_events.json', 'w', encoding='utf-8')
+        json.dump(afisha_data, events, ensure_ascii=False)
+    if typ == "article":
+        na = open('data/news_articles.json', encoding='utf-8')
+        f1 = na.read()
+        NA = json.loads(f1)
+        for k in NA:
+            if id == k['id']:
+                k['is_in_archive'] = True
+        na.close()
+        na = open('data/news_articles.json', 'w', encoding='utf-8')
+        json.dump(NA, na, ensure_ascii=False)
+    return redirect("/")
+
 
 if __name__ == '__main__':
     app.run(port=8080, host='127.0.0.1')
